Backend/Backend/requirements.txt
psycopg2==2.9.5 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
gunicorn==20.1.0  # https://github.com/benoitc/gunicorn
dj-database-url==1.0.0  #https://github.com/jazzband/dj-database-url
numpy==1.20.1
django-cors-headers==3.13.0

# Django
# ------------------------------------------------------------------------------
django==3.2.16  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client